fix(test): 環境変数上書きの再現で実効値を注入し .env 非定義環境の 500 を解消する (#6130)
forceKeyOverridden が $_SERVER にダミー値 'osvalue' を注入していたため,
.env に対象キーが無い環境（CI）で注入値がそのまま実効値になり,
twig.paths が存在しない app/template/osvalue を指して FilesystemLoader::addPath が
LoaderError を投げ, 全リクエストが 500 になっていた（PHPUnit 全 16 マトリクス fail）.
ローカルは .env に ECCUBE_TEMPLATE_CODE があり $_ENV が勝つため再現しなかった.

- 2 ファイルに重複していたヘルパを EnvOverrideTrait へ集約し, 注入値を引数化
- TemplateControllerTest は eccube.theme, SecurityControllerTest は
  eccube_admin_route の実効値を注入し, 環境差なく上書き状態のみを再現する
- 実効値でも SYMFONY_DOTENV_VARS 非掲載＋$_SERVER 有りの前提は変わらないため
  検出ロジックの検証意図は維持される
