+++ tests/Eccube/Tests/Web/Admin/Setting/System/SecurityControllerTest.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 
 namespace Eccube\Tests\Web\Admin\Setting\System;
 
+use Eccube\Tests\EnvOverrideTrait;
 use Eccube\Tests\Web\Admin\AbstractAdminWebTestCase;
 use PHPUnit\Framework\Attributes\Group;
 use Symfony\Component\HttpFoundation\Request;
@@ -22,6 +23,8 @@ use Symfony\Component\HttpFoundation\Request;
 #[Group('cache-clear')]
 final class SecurityControllerTest extends AbstractAdminWebTestCase
 {
+    use EnvOverrideTrait;
+
     protected $envFile;
 
     protected $env;
@@ -95,7 +98,7 @@ final class SecurityControllerTest extends AbstractAdminWebTestCase
      */
     public function testDisplayWarnsNamedKeyButKeepsButtonEnabledWhenSingleKeyOverridden(): void
     {
-        $this->forceKeyOverridden('ECCUBE_ADMIN_ROUTE', function (): void {
+        $this->forceKeyOverridden('ECCUBE_ADMIN_ROUTE', $this->currentAdminRoute(), function (): void {
             $crawler = $this->client->request(Request::METHOD_GET, $this->generateUrl('admin_setting_system_security'));
             $this->assertTrue($this->client->getResponse()->isSuccessful());
 
@@ -119,7 +122,7 @@ final class SecurityControllerTest extends AbstractAdminWebTestCase
         $session = $this->createSession($this->client);
         $formData = $this->createFormData();
 
-        $this->forceKeyOverridden('ECCUBE_ADMIN_ROUTE', function () use ($session, $formData): void {
+        $this->forceKeyOverridden('ECCUBE_ADMIN_ROUTE', $this->currentAdminRoute(), function () use ($session, $formData): void {
             $this->client->request(
                 Request::METHOD_POST,
                 $this->generateUrl('admin_setting_system_security'),
@@ -168,40 +171,15 @@ final class SecurityControllerTest extends AbstractAdminWebTestCase
     }
 
     /**
-     * 指定キーが「.env 由来でなく OS のプロセス環境変数で上書きされている」状態を
-     * 決定的に作り出してコールバックを実行し、終了後に $_SERVER を元へ復元する.
+     * 現在アプリケーションが使用している管理画面のルーティングプレフィックス.
      *
-     * テスト環境では bootEnv により .env のキーが SYMFONY_DOTENV_VARS に載るため
-     * putenv では上書き状態を再現できない（EnvFileService は SYMFONY_DOTENV_VARS を優先する）。
-     * そこで当該キーを SYMFONY_DOTENV_VARS から除外し、$_SERVER にプロセス環境変数値を注入する.
+     * ECCUBE_ADMIN_ROUTE の上書きを再現する際、実効値と異なる値を注入すると
+     * 管理画面の URL 自体が変わってしまうため、実効値をそのまま注入する
+     * （{@see EnvOverrideTrait::forceKeyOverridden()}）.
      */
-    private function forceKeyOverridden(string $key, callable $fn): void
+    private function currentAdminRoute(): string
     {
-        $sentinel = '__ECCUBE_TEST_UNSET__';
-        $origVars = \array_key_exists('SYMFONY_DOTENV_VARS', $_SERVER) ? $_SERVER['SYMFONY_DOTENV_VARS'] : $sentinel;
-        $origVal = \array_key_exists($key, $_SERVER) ? $_SERVER[$key] : $sentinel;
-
-        $vars = array_filter(
-            explode(',', (string) ($sentinel === $origVars ? '' : $origVars)),
-            fn ($k) => $k !== $key && '' !== $k
-        );
-        $_SERVER['SYMFONY_DOTENV_VARS'] = implode(',', $vars);
-        $_SERVER[$key] = 'osvalue';
-
-        try {
-            $fn();
-        } finally {
-            if ($sentinel === $origVars) {
-                unset($_SERVER['SYMFONY_DOTENV_VARS']);
-            } else {
-                $_SERVER['SYMFONY_DOTENV_VARS'] = $origVars;
-            }
-            if ($sentinel === $origVal) {
-                unset($_SERVER[$key]);
-            } else {
-                $_SERVER[$key] = $origVal;
-            }
-        }
+        return (string) static::getContainer()->getParameter('eccube_admin_route');
     }
 
     /**
